Allowed stock kernel modules to run (#7)
Modifications done by referring to other custom kernel. These are a complete list of changes made:
-Disabled Android Verified Boot
-Removed module signing and changed all configurations in out defconfig file to be compiled into the kernel instead of as modules
-Patched module.c to allow modules that do not match to still run.

These changes have been tested to allow Android to run on the phone with full functionality by allowing us to load unverified driver kernel modules which are required to interface with the hardware of the phone. This is by no means optimal, but as I have no desire to replace stock kernel modules, breaking functionality of the phone if we boot stock kernel, this will be the current way I intend to keep it.

These modifications are needed because stock kernel modules which are used as device drivers by Android are signed and verified. Stock kernel is also signed and verified, and we do not have the key. Hence, Android identifies us as unverified and does not load the modules, causing us to lose functionalities such as Wifi or audio.

We are keeping the stock kernel drivers because if they get replaced, the stock kernel will not load them without modifications. Since I want stock configuration present in the phone in the case of an irrecoverable failure due to my custom kernel (So if my custom kernel has issues a reboot brings us back to a known working state) , we will just make do with this for now.

Here are the 2 possible ways we can fix this:
1. Replace kernel modules once we are certain that my custom kernel is stable enough we don't need a backup. Not for anytime soon in my opinion since we really need the stability.
2. Add kernel modules into the init ramdisk loaded when Android boots. Unsure if this will work, but Android seems to offer a way to load kernel modules needed into a ramdisk: https://source.android.com/devices/bootloader/partitions/ramdisk-partitions

However, this is seems to be a non-problem for now. Given that our hardware works we likely do not have to do anything.

Unsure if these will completely fix this issue, but if they cause any issues this can be raised in newer issues. Notably for now, an error message pops up when booting Android which is not seen when booting custom kernels, even those compiled by my toolchain. This seems to be purely cosmetic, something perhaps we can just keep an eye out for but no need to spend time focusing on solving the issue.