Fix a compilation error picked up by Clang 11. (#13)
I will try to explain why we need to do this pull request. Firstly, why did this only break build in later Clang? I am really not sure. My best guess is that the fact Clang compiled this without throwing an error was a bug.

Why is this needed?
Well, it shouldn't, but a combination of flags made this a build breaking issue. They enabled -Wstrict-prototypes, which means all functions must declare their arguments explicitly, if it does not take an argument you need to specify by declaring it as foo(void). They then went to set strict prototypes as an error instead of a warning, so compilation will break because of this error. This is standard practice in Linux kernel development.

TLDR: If the kernel source has a function declared without explicity stating its arguments, even if it would be void, it will fail with the new Clang version.

FYI, for anyone still reading, the function literally just above it, which is very similar except for small changes gets it right. Oneplus kernel codebase is great isn't it?